add missing validation

+++ backend/controllers/deduction.controller.js
@@ -6,6 +6,11 @@ class DeductionController {
     createDeduction = async (req, res) => {
         try {
             const { employeeId, reason, amount, date } = req.body
+            if (!employeeId && !reason && !amount && !date) {
+                return res.status(400).json({
+                    error: 'Missing required fields.'
+                })
+            }
             const deduction = await this.DeductionService.createDeduction(employeeId, reason, amount, date)
             return res.status(200).json({
                 success: `Successfully created deduction. Reason: ${deduction.reason}. Amount: ${deduction.amount}. Date: ${deduction.date} `
@@ -21,6 +26,11 @@ class DeductionController {
     deleteDeduction = async (req, res) => {
         try {
             const { id } = req.params;
+            if (!id) {
+                return res.status(400).json({
+                    error: 'Missing required fields.'
+                })
+            }
             const deduction = await this.DeductionService.deleteDeduction(id)
             return res.status(200).json({
                 success: `Successfully removed deduction. id: ${deduction.id}`
@@ -37,6 +47,11 @@ class DeductionController {
         try {
             const { id } = req.params;
             const { employeeId, reason, amount, date } = req.body;
+            if (!id && !employeeId && !reason && !amount && !date) {
+                return res.status(400).json({
+                    error: 'Missing required fields.'
+                })
+            }
             const deduction = await this.DeductionService.updateDeduction(id, employeeId, reason, amount, date)
             return res.status(200).json({
                 success: `Successfully updated deduction. id: ${deduction.id}`
@@ -64,6 +79,11 @@ class DeductionController {
     getDeduction = async (req, res) => {
         try {
             const { id } = req.params
+            if (!id) {
+                return res.status(200).json({
+                    error: 'Missing required fields.'
+                })
+            }
             const deduction = await this.DeductionService.getDeduction(id)
             return res.status(200).json(deduction)
         } catch (err) {
